[ghstack-poisoned]

+++ backends/cadence/hifi/operators/tests/test_op_quantize_per_tensor.cpp
@@ -118,8 +118,8 @@ TEST_F(HiFiQuantizePerTensorTest, CheckSingleElementIntQuantize) {
   constexpr int64_t kQuantMin = std::numeric_limits<int32_t>::min();
   constexpr int64_t kQuantMax = std::numeric_limits<int32_t>::max();
   constexpr float kInputValue = 100.0f;
-  constexpr int32_t kExpectedOutputValue =
-      static_cast<int32_t>(kInputValue / kScale + kZeroPoint);
+  constexpr int32_t kExpectedOutputValue = static_cast<int32_t>(
+      static_cast<double>(kInputValue) / kScale + kZeroPoint);
 
   quantize_per_tensor_out(
       tf.make(sizes, {kInputValue}),
@@ -144,8 +144,8 @@ TEST_F(HiFiQuantizePerTensorTest, CheckSingleElementUInt16Quantize) {
   constexpr int64_t kQuantMin = std::numeric_limits<uint16_t>::min();
   constexpr int64_t kQuantMax = std::numeric_limits<uint16_t>::max();
   constexpr float kInputValue = 100.0f;
-  constexpr uint16_t kExpectedOutputValue =
-      static_cast<uint16_t>(kInputValue / kScale + kZeroPoint);
+  constexpr uint16_t kExpectedOutputValue = static_cast<uint16_t>(
+      static_cast<double>(kInputValue) / kScale + kZeroPoint);
 
   quantize_per_tensor_out(
       tf.make(sizes, {kInputValue}),
